client1/src/main/java/pl/discovery/client1/ConfigController.java
package pl.discovery.client1;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Arrays;
import java.util.List;

@RestController
@RequestMapping("/config")
public class ConfigController {

    private final ConfigService configService;

    public ConfigController(ConfigService configService) {
        this.configService = configService;
    }

    @GetMapping
    public List<Config> getConfig() {
        return configService.getConfigList();
    }
}

class Config {
    private String serviceName;
    private String url;

    public Config(String serviceName, String url) {
        this.serviceName = serviceName;
        this.url = url;
    }

    public Config() {

    }

    public String getServiceName() {
        return serviceName;
    }

    public void setServiceName(String serviceName) {
        this.serviceName = serviceName;
    }

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }
}